% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{add_growth_rates}
\alias{add_growth_rates}
\title{Function to calculate different growth rates}
\usage{
add_growth_rates(infections)
}
\arguments{
\item{infections}{The data table to which the growth rates should be added.}
}
\description{
Function to calculate different growth rates of the number of infections and add these to the main data set.
}
\details{
<TODO> The different growth rates should be described here.
}
