% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\alias{generate_map_restrictions}
\title{Functions for generating a mapping table of restriction ids to restriction labels}
\usage{
generate_map_restrictions(restrictions)
}
\arguments{
\item{restrictions}{A data table as returned by the function \code{read_restrictions}.}
}
\description{
On the basis of the restriction data table returned by \code{read_restrictions} a mapping table of restriction ids to restriction labels is generated.
}
