% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Functions for exporting the results}
\usage{
export_results(coefs, destination)
}
\arguments{
\item{coefs}{The data.table containing the restriction effects as returned by the function \code{fit_lme}.}

\item{destination}{The path to the destination file to which the results sould be written.}
}
\value{
Saves a csv file.
}
\description{
Saves the table containing the restriction effects to a csv file.
}
