% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{merge_data}
\alias{merge_data}
\title{Function for merging infection and restriction data}
\usage{
merge_data(infections, restrictions, border_closures, map)
}
\arguments{
\item{infections}{Return value from function \code{read_infections}.}

\item{restrictions}{Return value from function \code{read_restrictions}.}

\item{border_closures}{Return value from function \code{read_border_closures}.}

\item{map}{Return value from function \code{read_map_bundesland}.}
}
\description{
Function for merging infection and restriction data on the basis of the return values of the functions \code{read_infections}, \code{read_restrictions}, \code{read_border_closures} and \code{read_map_bundesland}.
}
