% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_lme}
\alias{fit_lme}
\title{Function for estimatind the effects of restrictions on the (smoothed) lograte of the the number of infections}
\usage{
fit_lme(data, map_restrictions, cutoff_date)
}
\arguments{
\item{data}{The data table returned from the function \code{add_growth_rates}.}

\item{map_restrictions}{The mapping table returned by the function \code{generate_map_restrictions}.}

\item{cutoff_date}{The cut-off data at which the effects should be estimated (only used here as a label in the resulting table of estimates).}
}
\value{
The fitted model and the coefficient estimations for the fixed effects are returned as a data.table.
}
\description{
Fits a linear mixed equations model effects and returns the parameter estimations for the resitrictions.
}
\details{
The model adds a random intercept and a random effect for the variable 'day' (days since first infection) per federal state.  In addition, fixed effects for the number of days since the effectuation of each restriction are added.
}
