% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\alias{read_map_bundesland}
\alias{read_infections}
\alias{read_border_closures}
\alias{read_restrictions}
\title{Functions for reading raw data}
\usage{
read_map_bundesland(file)

read_infections(file, cutoff_date)

read_border_closures(file, cutoff_date)

read_restrictions(file, border_data, cutoff_date)
}
\arguments{
\item{file}{Path to csv file containing the data.  If no file path is specified, data are lodad via a Statworx API.}

\item{cutoff_date}{The cut-off data at which the effects should be estimated. (infection data and restrictions imposed after this date are not used).}

\item{border_data}{A \code{data.table} containing the border closure data returned by the function \code{read_border_closures}.}
}
\value{
An objet of class \code{data.table}.
}
\description{
Functions for reading data on infections, border closures and further restrictions.
}
\details{
\code{read_map_bundesland} reads data for creating a table for mapping federal state names to their two-letter codes.
}
