% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{remove_webasto}
\alias{remove_webasto}
\title{Functions for removing infection data pertaining to the company Webasto in Bavaria}
\usage{
remove_webasto(infections, start_wave2 = "2020-02-28")
}
\arguments{
\item{infections}{The data table from which to remove data.}

\item{start_wave2}{The date of the first infection not associated with Webasto.}
}
\description{
In Bavaria there appears to habe been two waves of infections and all data pertaining to the first one (associated with the company Webasto) are removed.
}
\details{
The infection numbers before the date specified in start_wave2 are removed and the remaining are set back accordingly.
}
